Метод для тестов updateCommonModule перенесен в отдельный класс
