+++ tests/com.e1c.v8codestyle.md.itests/src/com/e1c/v8codestyle/md/itests/CommonModuleNameClientTest.java
@@ -16,22 +16,12 @@ package com.e1c.v8codestyle.md.itests;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 
-import java.util.Map;
-import java.util.Map.Entry;
-
-import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.emf.ecore.EStructuralFeature;
 import org.junit.Test;
 
-import com._1c.g5.v8.bm.core.IBmObject;
-import com._1c.g5.v8.bm.core.IBmTransaction;
-import com._1c.g5.v8.bm.integration.AbstractBmTask;
-import com._1c.g5.v8.bm.integration.IBmModel;
 import com._1c.g5.v8.dt.core.platform.IDtProject;
-import com._1c.g5.v8.dt.metadata.mdclass.CommonModule;
 import com._1c.g5.v8.dt.metadata.mdclass.ReturnValuesReuse;
 import com._1c.g5.v8.dt.validation.marker.Marker;
-import com.e1c.g5.v8.dt.testing.check.CheckTestBase;
+import com.e1c.v8codestyle.internal.md.itests.CheckMd;
 import com.e1c.v8codestyle.md.check.CommonModuleNameClient;
 import com.e1c.v8codestyle.md.check.CommonModuleType;
 
@@ -42,7 +32,7 @@ import com.e1c.v8codestyle.md.check.CommonModuleType;
  *
  */
 public class CommonModuleNameClientTest
-    extends CheckTestBase
+    extends CheckMd
 {
 
     private static final String CHECK_ID = "common-module-name-client";
@@ -125,47 +115,6 @@ public class CommonModuleNameClientTest
         long id = getTopObjectIdByFqn(MODULE_DEFAULT_FQN, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
         assertNull(marker);
-    }
-
-    private void updateCommonModule(IDtProject dtProject, String fqn, Map<EStructuralFeature, Boolean> types,
-        ReturnValuesReuse returnValueReuse, String newFqn)
-    {
-        IBmModel model = bmModelManager.getModel(dtProject);
-        model.execute(new AbstractBmTask<Void>("change type")
-        {
-            @Override
-            public Void execute(IBmTransaction transaction, IProgressMonitor monitor)
-            {
-                IBmObject object = transaction.getTopObjectByFqn(fqn);
-
-                for (Entry<EStructuralFeature, Boolean> entry : types.entrySet())
-                {
-                    object.eSet(entry.getKey(), entry.getValue());
-                }
-
-                if (!(object instanceof CommonModule))
-                {
-                    return null;
-                }
-
-                CommonModule module = (CommonModule)object;
-
-                module.setReturnValuesReuse(returnValueReuse);
-
-                if (newFqn != null)
-                {
-                    String[] fqnArray = newFqn.split("[.]");
-                    if (fqnArray.length == 2)
-                    {
-                        module.setName(fqnArray[1]);
-                        transaction.updateTopObjectFqn(object, newFqn);
-                    }
-                }
-
-                return null;
-            }
-        });
-        waitForDD(dtProject);
     }
 
 }
