+++ tests/com.e1c.v8codestyle.md.itests/src/com/e1c/v8codestyle/md/itests/CommonModuleTypeTest.java
@@ -16,20 +16,11 @@ package com.e1c.v8codestyle.md.itests;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 
-import java.util.Map;
-import java.util.Map.Entry;
-
-import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.emf.ecore.EStructuralFeature;
 import org.junit.Test;
 
-import com._1c.g5.v8.bm.core.IBmObject;
-import com._1c.g5.v8.bm.core.IBmTransaction;
-import com._1c.g5.v8.bm.integration.AbstractBmTask;
-import com._1c.g5.v8.bm.integration.IBmModel;
 import com._1c.g5.v8.dt.core.platform.IDtProject;
 import com._1c.g5.v8.dt.validation.marker.Marker;
-import com.e1c.g5.v8.dt.testing.check.CheckTestBase;
+import com.e1c.v8codestyle.internal.md.itests.CheckMd;
 import com.e1c.v8codestyle.md.check.CommonModuleType;
 
 /**
@@ -39,7 +30,7 @@ import com.e1c.v8codestyle.md.check.CommonModuleType;
  *
  */
 public class CommonModuleTypeTest
-    extends CheckTestBase
+    extends CheckMd
 {
 
     private static final String CHECK_ID = "common-module-type";
@@ -75,7 +66,7 @@ public class CommonModuleTypeTest
 
         String fqn = "CommonModule.Common";
 
-        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_SERVER);
+        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_SERVER, null, null);
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
         assertNull(marker);
@@ -110,7 +101,7 @@ public class CommonModuleTypeTest
 
         String fqn = "CommonModule.CommonServerCall";
 
-        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_SERVER_CALL);
+        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_SERVER_CALL, null, null);
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
         assertNull(marker);
@@ -145,7 +136,7 @@ public class CommonModuleTypeTest
 
         String fqn = "CommonModule.CommonClient";
 
-        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_CLIENT);
+        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_CLIENT, null, null);
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
         assertNull(marker);
@@ -180,7 +171,7 @@ public class CommonModuleTypeTest
 
         String fqn = "CommonModule.CommonServerClient";
 
-        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_CLIENT_SERVER);
+        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_CLIENT_SERVER, null, null);
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
         assertNull(marker);
@@ -215,7 +206,7 @@ public class CommonModuleTypeTest
 
         String fqn = "CommonModule.CommonServerGlobal";
 
-        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_SERVER_GLOBAL);
+        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_SERVER_GLOBAL, null, null);
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
         assertNull(marker);
@@ -250,28 +241,10 @@ public class CommonModuleTypeTest
 
         String fqn = "CommonModule.CommonClientGlobal";
 
-        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_CLIENT_GLOBAL);
+        updateCommonModule(dtProject, fqn, CommonModuleType.TYPE_CLIENT_GLOBAL, null, null);
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
         assertNull(marker);
     }
 
-    private void updateCommonModule(IDtProject dtProject, String fqn, Map<EStructuralFeature, Boolean> types)
-    {
-        IBmModel model = bmModelManager.getModel(dtProject);
-        model.execute(new AbstractBmTask<Void>("change type")
-        {
-            @Override
-            public Void execute(IBmTransaction transaction, IProgressMonitor monitor)
-            {
-                IBmObject object = transaction.getTopObjectByFqn(fqn);
-                for (Entry<EStructuralFeature, Boolean> entry : types.entrySet())
-                {
-                    object.eSet(entry.getKey(), entry.getValue());
-                }
-                return null;
-            }
-        });
-        waitForDD(dtProject);
-    }
 }
